+++ src/com/intellij/idea/plugin/hybris/system/cockpitng/lang/folding/CngConfigFoldingBuilder.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -37,12 +37,13 @@ import com.intellij.openapi.project.DumbAware
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiErrorElement
 import com.intellij.psi.SyntaxTraverser
-import com.intellij.psi.util.PsiElementFilter
 import com.intellij.psi.xml.XmlFile
 import com.intellij.psi.xml.XmlTag
 import com.intellij.util.xml.DomManager
 
 class CngConfigFoldingBuilder : FoldingBuilderEx(), DumbAware {
+
+    private val filter = CngConfigFilter()
 
     override fun buildFoldRegions(root: PsiElement, document: Document, quick: Boolean): Array<FoldingDescriptor> {
         if (!HybrisProjectSettingsComponent.getInstance(root.project).isHybrisProject()) return emptyArray()
@@ -51,7 +52,7 @@ class CngConfigFoldingBuilder : FoldingBuilderEx(), DumbAware {
             ?: return emptyArray()
 
         return SyntaxTraverser.psiTraverser(root)
-            .filter { CngConfigFilter().isAccepted(it) }
+            .filter { filter.isAccepted(it) }
             .mapNotNull {
                 if (it is PsiErrorElement || it.textRange.isEmpty) return@mapNotNull null
                 FoldingDescriptor(it.node, it.textRange, FoldingGroup.newGroup(GROUP_NAME))
@@ -99,22 +100,6 @@ class CngConfigFoldingBuilder : FoldingBuilderEx(), DumbAware {
         }
 
         else -> false
-    }
-
-    class CngConfigFilter : PsiElementFilter {
-
-        override fun isAccepted(element: PsiElement) = when (element) {
-            is XmlTag -> when (element.localName) {
-                FieldList.FIELD,
-                Section.ATTRIBUTE,
-                ExplorerTree.TYPE_NODE,
-                ListView.COLUMN -> true
-
-                else -> false
-            }
-
-            else -> false
-        }
     }
 
     companion object {