+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## [2023.2.4]
+
+### `FlexibleSearch` enhancements
+- Improved language injection into Java files [#515](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/515)
+
+### Features
+- Improved performance of the Item attribute resolution [#516](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/516)
+- Added possibility to skip non-existing source directories during project import [#511](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/511)
+
+### Fixes
+- Missing navigation to Bean Enum declaration from java class [#517](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/517)
+- Fixed deadlock when ItemType name equals to its extends [#513](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/513)
+- Refactored `ImportProjectProgressModalWindow` so that is calls so that project state retrieved only once [#512](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/512)
+- Changed generated `*.iml` file name when grouping is not selected so file name does not start with a dot [#512](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/512)
+
+### Other
+- Updated Kotlin to 1.9.0
+- Updated Gradle plugin to 1.15.0
+- Updated Gradle to 8.2.1
+
+### Deprecated
+- Updated usage of the Diagram API
+
 ## [2023.2.3]
 
 ### Features
@@ -562,7 +585,7 @@
 - Extended/refactored type system inspection rules
 - Merged ToolsWindows "Hybris" (as Remote Instances) and "Hybris Console" (as Consoles) altogether with "Type system" into "Hybris" as a single entry point for all [y] related actions
 
-### Deprecation
+### Deprecated
 - Decreased usage of the Deprecated API
 
 ## [2022.2]